Make sure a DRAWSEGMENT is actually a straight line
This change will stop the panelizer from making holes in circular or arc
board outline. However this would not work correctly anyway so this
change just stops the panelizer from breaking the board outline ever
more.

More special cases should be added for handling the board outline when
it is made of circles, arcs or when the tab might not intersect with
just a single straight line.